refactor(core): move pure engine files to packages/core
