+++ packages/core/src/ir/printer.ts
