+++ packages/core/src/utils/ignore-pragma.ts
