+++ packages/core/src/utils/oxc-parse.ts
